Move all the legacy components to a separate section
