+++ ts/features/design-system/core/DSButtons.tsx
@@ -112,28 +112,6 @@ export const DSButtons = () => {
         IconButtonContained (Icebox)
       </H2>
       {renderIconButtonContained(isExperimental)}
-
-      <VSpacer size={48} />
-
-      <H2
-        color={theme["textHeading-default"]}
-        weight={"SemiBold"}
-        style={{ marginBottom: 16 }}
-      >
-        Block Buttons (not NativeBase)
-      </H2>
-      {renderBlockButtons()}
-
-      <VSpacer size={48} />
-
-      <H2
-        color={theme["textHeading-default"]}
-        weight={"SemiBold"}
-        style={{ marginBottom: 16 }}
-      >
-        Specific buttons
-      </H2>
-      {renderSpecificButtons()}
     </DesignSystemScreen>
   );
 };
@@ -992,138 +970,6 @@ const renderIconButtonContained = (isExperimental: boolean) => (
         </View>
       </DSComponentViewerBox>
     </View>
-  </>
-);
-
-const renderBlockButtons = () => (
-  <>
-    <BlockButtons
-      type="SingleButton"
-      primary={{
-        type: "Solid",
-        buttonProps: {
-          color: "primary",
-          accessibilityLabel: "primary button",
-          onPress: onButtonPress,
-          label: "Primary button"
-        }
-      }}
-    />
-    <VSpacer size={16} />
-    <BlockButtons
-      type="TwoButtonsInlineThird"
-      primary={{
-        type: "Outline",
-        buttonProps: {
-          color: "primary",
-          accessibilityLabel: "Left button",
-          onPress: onButtonPress,
-          label: "Left button"
-        }
-      }}
-      secondary={{
-        type: "Solid",
-        buttonProps: {
-          color: "primary",
-          accessibilityLabel: "Right button",
-          onPress: onButtonPress,
-          label: "Right button"
-        }
-      }}
-    />
-    <VSpacer size={16} />
-    <BlockButtons
-      type="TwoButtonsInlineHalf"
-      primary={{
-        type: "Outline",
-        buttonProps: {
-          color: "primary",
-          accessibilityLabel: "Left button",
-          onPress: onButtonPress,
-          label: "Left button"
-        }
-      }}
-      secondary={{
-        type: "Solid",
-        buttonProps: {
-          color: "primary",
-          accessibilityLabel: "Right button",
-          onPress: onButtonPress,
-          label: "Right button"
-        }
-      }}
-    />
-    <VSpacer size={16} />
-    <BlockButtons
-      type="TwoButtonsInlineThirdInverted"
-      primary={{
-        type: "Outline",
-        buttonProps: {
-          color: "primary",
-          accessibilityLabel: "Left button",
-          onPress: onButtonPress,
-          label: "Left button"
-        }
-      }}
-      secondary={{
-        type: "Solid",
-        buttonProps: {
-          color: "primary",
-          accessibilityLabel: "Right button",
-          onPress: onButtonPress,
-          label: "Right button"
-        }
-      }}
-    />
-  </>
-);
-
-const renderSpecificButtons = () => (
-  <>
-    <DSComponentViewerBox name="PaymentButton">
-      <PaymentButton
-        amount={9999999999}
-        noticeNumber={"123112312312321321" as PaymentNoticeNumber}
-        organizationFiscalCode={"46545" as OrganizationFiscalCode}
-      />
-    </DSComponentViewerBox>
-    <DSComponentViewerBox name="CalendarEventButton (using new ButtonOutline)">
-      <ButtonOutline
-        accessibilityLabel="Tap to trigger test alert"
-        label={"Aggiungi promemoria"}
-        icon="add"
-        onPress={onButtonPress}
-      />
-
-      <VSpacer size={16} />
-
-      <ButtonOutline
-        accessibilityLabel="Tap to trigger test alert"
-        label={"Aggiunto"}
-        icon="checkTickBig"
-        onPress={onButtonPress}
-      />
-    </DSComponentViewerBox>
-    <DSComponentViewerBox name="CopyButtonComponent">
-      <CopyButtonComponent textToCopy={"Copied text by CopyButton"} />
-    </DSComponentViewerBox>
-    <DSComponentViewerBox name="Login buttons">
-      <ButtonSolid
-        fullWidth
-        accessibilityLabel="Tap to trigger test alert"
-        label={"Entra con SPID"}
-        icon="profile"
-        onPress={onButtonPress}
-      />
-      <VSpacer size={8} />
-      <ButtonSolid
-        fullWidth
-        accessibilityLabel="Tap to trigger test alert"
-        label={"Entra con CIE"}
-        icon="cie"
-        onPress={onButtonPress}
-      />
-    </DSComponentViewerBox>
   </>
 );
