+++ ts/features/design-system/navigation/params.ts
@@ -46,4 +46,12 @@ export type DesignSystemParamsList = {
   [DESIGN_SYSTEM_ROUTES.DEBUG.FOOTER_ACTIONS_NOT_FIXED.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.DEBUG.SCREEN_END_MARGIN.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.LEGACY.PICTOGRAMS.route]: undefined;
+  [DESIGN_SYSTEM_ROUTES.LEGACY.BUTTONS.route]: undefined;
+  [DESIGN_SYSTEM_ROUTES.LEGACY.TEXT_FIELDS.route]: undefined;
+  [DESIGN_SYSTEM_ROUTES.LEGACY.LIST_ITEMS.route]: undefined;
+  [DESIGN_SYSTEM_ROUTES.LEGACY.BADGES.route]: undefined;
+  [DESIGN_SYSTEM_ROUTES.LEGACY.SELECTION.route]: undefined;
+  [DESIGN_SYSTEM_ROUTES.LEGACY.ACCORDION.route]: undefined;
+  [DESIGN_SYSTEM_ROUTES.LEGACY.ALERT.route]: undefined;
+  [DESIGN_SYSTEM_ROUTES.LEGACY.ADVICE.route]: undefined;
 };
