+++ ts/features/design-system/navigation/routes.ts
@@ -114,6 +114,38 @@ const DESIGN_SYSTEM_ROUTES = {
     PICTOGRAMS: {
       route: "DESIGN_SYSTEM_LEGACY_PICTOGRAMS",
       title: "Pictograms"
+    },
+    BUTTONS: {
+      route: "DESIGN_SYSTEM_LEGACY_BUTTONS",
+      title: "Buttons"
+    },
+    TEXT_FIELDS: {
+      route: "DESIGN_SYSTEM_LEGACY_TEXT_FIELDS",
+      title: "Text Fields"
+    },
+    LIST_ITEMS: {
+      route: "DESIGN_SYSTEM_LEGACY_LIST_ITEMS",
+      title: "List Items"
+    },
+    BADGES: {
+      route: "DESIGN_SYSTEM_LEGACY_BADGES",
+      title: "Badges"
+    },
+    SELECTION: {
+      route: "DESIGN_SYSTEM_LEGACY_SELECTION",
+      title: "Selection"
+    },
+    ACCORDION: {
+      route: "DESIGN_SYSTEM_LEGACY_ACCORDION",
+      title: "Accordion"
+    },
+    ALERT: {
+      route: "DESIGN_SYSTEM_LEGACY_ALERT",
+      title: "Alert"
+    },
+    ADVICE: {
+      route: "DESIGN_SYSTEM_LEGACY_ADVICE",
+      title: "Advice & Banners"
     }
   } as const
 } as const;
